Removed unnecessary files when everything is build on dev machine

+++ Scripts/vercel-build.sh
@@ -1,5 +1,5 @@
 #! /bin/bash
-# Time-stamp: "2021-01-28 09:43:18 queinnec"
+# Time-stamp: "2021-01-30 09:35:42 queinnec"
 
 # Build the P server on Vercel.
 # api/p.js should already exist to be taken into account.
@@ -22,11 +22,13 @@
 #   manifest.json
 #   service-worker-index.html
 #   service-worker.js
-# So, static/ is exported
+# So, the final ./export/ directory is made with
+#     the content of static/
 #     __sapper__/export is moved into static/client
-#     api/ is moved into static/
+#     and the api/ directory is moved into static/
 
 DEBUG=false
+DEBUG=true
 
 showls () {
     if $DEBUG
@@ -49,12 +51,19 @@ then
     exec /bin/true
 fi
 
+BUILD_EXPORT=true
+BUILD_SERVER=true
+
 # The webapp and its serverless functions are completely built on the
 # dev machine so nothing else to be done on Vercel.
 if [ -d export ]
 then
-   echo "*** The entire webapp and its serverless functions are ready"
-   exec /bin/true
+    BUILD_EXPORT=false
+fi
+
+if [ "$(wc -l < api/p.js)" -gt 1000 ]
+then
+    BUILD_SERVER=false
 fi
 
 # Inquire context:
@@ -70,34 +79,44 @@ then
     printenv
 fi
 
-echo "*** Building the p.codegradx.org static server..."
-# export needs the original server.js with a listen method in order to
-# crawl the site after generation.
-cp -pf src/nativeserver.js src/server.js
-
-npm run export
-showls `pwd`/
-showls __sapper__/
-showls __sapper__/export/
-
-if [ ! -f __sapper__/export/index.html ]
+if $BUILD_EXPORT
 then
-    echo "!!!!! Crawling phase probably missing!" >&2
-    exit 1
+    echo "*** Building the p.codegradx.org static server..."
+    # export needs the original server.js with a listen method in
+    # order to crawl the site after generation.
+    cp -pf src/nativeserver.js src/server.js
+    npm run export
+    showls `pwd`/
+    showls __sapper__/
+    showls __sapper__/export/
+
+    if [ ! -f __sapper__/export/index.html ]
+    then
+        echo "!!!!! Crawling phase probably missing!" >&2
+        exit 1
+    fi
+    # Make sure that ./export is the 'output directory' in the project settings:
+    mv __sapper__/export .
+    # Remove that directory in order to be able to build the dynamic server:
+    rm -rf __sapper__/build
+else
+    echo "*** Don't rebuild the ./export/ directory!"
 fi
-# Make sure that ./export is the 'output directory' in the project settings:
-mv __sapper__/export .
-rm -rf __sapper__/build
 
-echo "*** Building the p.codegradx.org dynamic server..."
-# Update src/server.js in order to be invoked as a serverless function.
-cp -pf src/pserver.js src/server.js
-npm run build
-tr -d '\r' < __sapper__/build/server/server.js > api/p.js
-showls api/
+if $BUILD_SERVER
+then
+    echo "*** Building the p.codegradx.org dynamic server..."
+    # Update src/server.js in order to be invoked as a serverless function.
+    cp -pf src/pserver.js src/server.js
+    npm run build
+    tr -d '\r' < __sapper__/build/server/server.js > api/p.js
+    showls api/
 
-# Restaure back the original file:
-mv -f src/nativeserver.js src/server.js
+    # Restaure back the original file:
+    mv -f src/nativeserver.js src/server.js
+else
+    echo "*** Don't rebuild the api/p.js dynamic server!"
+fi
 
 echo "*** end of vercel-build.sh"
